Working version of Big Guy, general lighting dimmed
