--- Assets/AnimationEventHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4074f97e1027c6046a1efe10be705b80
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
