+++ Assets/Animations/Enemies/Big Guy/big_guy_attack_up.anim
@@ -100,6 +100,52 @@ AnimationClip:
     path: 
     classID: 114
     script: {fileID: 11500000, guid: 8c6ff944d3f8f994b811165e5e40aa36, type: 3}
+  - curve:
+      serializedVersion: 2
+      m_Curve:
+      - serializedVersion: 3
+        time: 0
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1
+        value: 1
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1.0833334
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1.8333334
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      m_PreInfinity: 2
+      m_PostInfinity: 2
+      m_RotationOrder: 4
+    attribute: isWindBlasting
+    path: FistHitbox/UpHB
+    classID: 114
+    script: {fileID: 11500000, guid: 952c4cebcecd7324683585e6eda90cb4, type: 3}
   m_PPtrCurves:
   - curve:
     - time: 0
@@ -142,6 +188,13 @@ AnimationClip:
       path: 0
       attribute: 60754182
       script: {fileID: 11500000, guid: 8c6ff944d3f8f994b811165e5e40aa36, type: 3}
+      typeID: 114
+      customType: 0
+      isPPtrCurve: 0
+    - serializedVersion: 2
+      path: 2459823563
+      attribute: 2287089382
+      script: {fileID: 11500000, guid: 952c4cebcecd7324683585e6eda90cb4, type: 3}
       typeID: 114
       customType: 0
       isPPtrCurve: 0
@@ -266,6 +319,52 @@ AnimationClip:
     path: 
     classID: 114
     script: {fileID: 11500000, guid: 8c6ff944d3f8f994b811165e5e40aa36, type: 3}
+  - curve:
+      serializedVersion: 2
+      m_Curve:
+      - serializedVersion: 3
+        time: 0
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1
+        value: 1
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1.0833334
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      - serializedVersion: 3
+        time: 1.8333334
+        value: 0
+        inSlope: Infinity
+        outSlope: Infinity
+        tangentMode: 103
+        weightedMode: 0
+        inWeight: 0
+        outWeight: 0
+      m_PreInfinity: 2
+      m_PostInfinity: 2
+      m_RotationOrder: 4
+    attribute: isWindBlasting
+    path: FistHitbox/UpHB
+    classID: 114
+    script: {fileID: 11500000, guid: 952c4cebcecd7324683585e6eda90cb4, type: 3}
   m_EulerEditorCurves: []
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
