--- Assets/BigGuyMelee.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 952c4cebcecd7324683585e6eda90cb4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
