--- Assets/CameraZoomController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a602c3ac51973c44d93e8421b0956b5e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
