--- Assets/EnemyBarrier.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac15d2407a606ba4686185dc4dfd4da1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
