--- Assets/HealthBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20500bb8fbbb3b94e8ac5f0cb76cde1f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
