+++ Assets/Prefabs/Spells/Windblast_Spell.prefab
@@ -4911,7 +4911,7 @@ GameObject:
   - component: {fileID: 7995059032508458879}
   m_Layer: 0
   m_Name: Windblast_Spell
-  m_TagString: Spell
+  m_TagString: WindBlast
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -4970,12 +4970,12 @@ MonoBehaviour:
   spell: {fileID: 11400000, guid: cb799c27f5fb04949a73f3d9c4bbd993, type: 2}
   collider: {fileID: 0}
   rb: {fileID: 7410404838220746693}
-  rotate: {x: 0, y: 0, z: 180}
+  rotate: {x: 0, y: 0, z: -90}
   spellActive: 1
   isChair: 0
   isWind: 1
   isPiplup: 0
-  direction: {x: 0, y: -1}
+  direction: {x: 1, y: 0}
   damage: 0
 --- !u!61 &9097417068368567135
 BoxCollider2D:
