--- Assets/RoomControlA2Old.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8850ce320e038584494734560d183b56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
