+++ Assets/Scripts/AnimationEventHandler.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AnimationEventHandler : MonoBehaviour
+{
+    
+    
+}
