+++ Assets/Scripts/CameraZoomController.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CameraZoomController : MonoBehaviour
+{
+    [SerializeField] private UnityEngine.Experimental.Rendering.Universal.PixelPerfectCamera ppc;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void ChangeAssetPixels(int num) {
+        ppc.assetsPPU = num;
+    }
+}
