+++ Assets/Scripts/EnemyBarrier.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyBarrier : MonoBehaviour
+{
+    private void OnCollisionEnter2D(Collision2D other) {
+        if (other.gameObject.tag == "Player") {
+            Destroy(this.gameObject);
+        }
+    }
+}
