+++ Assets/Scripts/IceSpell_Behavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 576cecbb2a863294ca0332dde34d8f23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
