+++ Assets/Scripts/MainExit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 612699b627098784f88673164fce4021
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
