+++ Assets/Scripts/TextOverhead.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddbbe07760c9e0a44b6a48888d121c77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
